django/requirements.txt
Babel==2.9.0
caluma==7.2.0
dataclasses-json==0.5.2
Django==2.2.17
django-clamd==0.4.0
django-environ==0.4.5
django-excel==0.0.10
django-extra-fields==3.0.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework-xml==2.0.0
django-reversion==3.0.8
docxtpl==0.11.2
drf-yasg==1.20.0
Jinja2==2.11.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pyjexl==0.2.3
PyJWT==1.7.1
pyproj==3.0.0.post1
python-dateutil==2.8.1
python-keycloak==0.22.0
python-memcached==1.59
PyXB==1.2.6
requests==2.25.0
sorl-thumbnail==12.7.0
tabulate==0.8.7
uWSGI==2.0.19.1
xmlschema==1.3.1
manabi @ git+https://github.com/ganwell/manabi
